feat: extract ingestor Angular Lib to Node Lib
