+++ libs/ingestor/.eslintrc.json
@@ -1,0 +1,1 @@
+{ "extends": "../../.eslintrc.json", "ignorePatterns": ["!**/*"], "rules": {} }
