+++ libs/ingestor/README.md
@@ -1,0 +1,7 @@
+# ingestor
+
+This library was generated with [Nx](https://nx.dev).
+
+## Running unit tests
+
+Run `nx test ingestor` to execute the unit tests via [Jest](https://jestjs.io).
