+++ libs/ingestor/package.json
@@ -1,0 +1,4 @@
+{
+  "name": "@xlayers/ingestor",
+  "version": "0.0.1"
+}
