+++ libs/ingestor/src/lib/ingestor.spec.ts
@@ -1,0 +1,7 @@
+import { ingestor } from './ingestor';
+
+describe('ingestor', () => {
+  it('should work', () => {
+    expect(ingestor()).toEqual('ingestor');
+  });
+});
