+++ package.json
@@ -74,6 +74,7 @@
     "@ngxs/logger-plugin": "^3.6.2",
     "@nrwl/cli": "11.0.15",
     "@nrwl/cypress": "11.0.15",
+    "@nrwl/eslint-plugin-nx": "11.0.15",
     "@nrwl/jest": "11.0.15",
     "@nrwl/node": "11.0.15",
     "@nrwl/tao": "11.0.15",
@@ -86,6 +87,8 @@
     "@types/jest": "26.0.8",
     "@types/node": "12.12.38",
     "@types/sketchapp": "^56.0.0",
+    "@typescript-eslint/eslint-plugin": "4.3.0",
+    "@typescript-eslint/parser": "4.3.0",
     "@xlayers/version-stamp": "^1.0.0",
     "adm-zip": ">=0.4.11",
     "codelyzer": "6.0.1",
@@ -94,6 +97,7 @@
     "cypress-file-upload": "^4.0.7",
     "dotenv": "6.2.0",
     "eslint": "6.8.0",
+    "eslint-config-prettier": "6.0.0",
     "husky": "^4.2.5",
     "jest": "26.2.2",
     "jest-preset-angular": "8.3.1",
